+++ src/main/java/com/github/thedeathlycow/thermoo/api/core/v1/event/LivingEntityTemperatureTickEvents.java
@@ -1,4 +1,4 @@
-package com.github.thedeathlycow.thermoo.api.temperature.event;
+package com.github.thedeathlycow.thermoo.api.core.v1.event;
 
 import net.fabricmc.fabric.api.event.Event;
 import net.fabricmc.fabric.api.event.EventFactory;
