+++ src/main/java/com/github/thedeathlycow/thermoo/api/core/v1/event/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains classes related to temperature events
+ */
+package com.github.thedeathlycow.thermoo.api.core.v1.event;
