--- src/main/java/com/github/thedeathlycow/thermoo/api/temperature/event/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Contains classes related to temperature events
- */
-package com.github.thedeathlycow.thermoo.api.temperature.event;
