+++ src/main/java/com/github/thedeathlycow/thermoo/impl/LivingEntityTickUtil.java
@@ -3,9 +3,9 @@ package com.github.thedeathlycow.thermoo.impl;
 import com.github.thedeathlycow.thermoo.api.environment.EnvironmentLookup;
 import com.github.thedeathlycow.thermoo.api.core.v1.HeatingMode;
 import com.github.thedeathlycow.thermoo.api.core.v1.HeatingModes;
-import com.github.thedeathlycow.thermoo.api.temperature.event.EnvironmentTickContext;
-import com.github.thedeathlycow.thermoo.api.temperature.event.LivingEntitySoakingTickEvents;
-import com.github.thedeathlycow.thermoo.api.temperature.event.LivingEntityTemperatureTickEvents;
+import com.github.thedeathlycow.thermoo.api.core.v1.event.EnvironmentTickContext;
+import com.github.thedeathlycow.thermoo.api.core.v1.event.LivingEntitySoakingTickEvents;
+import com.github.thedeathlycow.thermoo.api.core.v1.event.LivingEntityTemperatureTickEvents;
 import com.github.thedeathlycow.thermoo.impl.component.ThermooComponents;
 import com.github.thedeathlycow.thermoo.impl.environment.EnvironmentTickContextImpl;
 import com.github.thedeathlycow.thermoo.impl.environment.ServerPlayerTickUtil;
